Migrate logFieldsErrors
