+++ packages/cli-lib/errorHandlers.js
@@ -15,6 +15,8 @@ const {
   parseValidationErrors,
 } = require('./errorHandlers/apiErrors');
 
+const { logFieldJsErrors } = require('./errorHandlers/fieldsJsErrors');
+
 module.exports = {
   ErrorContext,
   ApiErrorContext,
@@ -26,4 +28,5 @@ module.exports = {
   logApiUploadErrorInstance,
   logFileSystemErrorInstance,
   logServerlessFunctionApiErrorInstance,
+  logFieldJsErrors,
 };
