+++ packages/cli-lib/errorHandlers/fieldsJsErrors.js
@@ -1,0 +1,29 @@
+const { i18n } = require('@hubspot/cli-lib/lib/lang');
+const { logger } = require('../logger');
+const { getExt } = require('../path');
+const i18nKey = 'cli.commands.upload';
+
+function logFieldJsErrors(e, filePath) {
+  if (e instanceof SyntaxError) {
+    const ext = getExt(filePath);
+    if (ext == 'js') {
+      logger.error(i18n(`${i18nKey}.errors.jsSyntaxError`, { js: filePath }));
+    }
+  }
+  if (e.code === 'ENOENT') {
+    logger.error(
+      i18n(`${i18nKey}.errors.invalidPath`, {
+        path: filePath,
+      })
+    );
+  }
+  if (e.code === 'MODULE_NOT_FOUND') {
+    logger.error(
+      i18n(`${i18nKey}.errors.jsSyntaxError`, {
+        path: filePath,
+      })
+    );
+  }
+}
+
+module.exports = logFieldJsErrors;
